+++ memoria/include/procesar_conexion.h
@@ -17,7 +17,7 @@
 
     void procesar_conexion_cpu_memoria(int socket_cliente);
 
-    void enviar_config_a_cpu(int socket_cliente, t_log* logger, uint8_t paginas_por_tabla, uint8_t tamanio_pagina);
+    void enviar_config_a_cpu(int socket_cliente, t_log* logger, uint32_t paginas_por_tabla, uint32_t tamanio_pagina);
     
     void enviar_boludeces_a_cpu(int socket_cliente, int32_t nro_tabla_segundo_nivel);
